+++ backend/src/routes/index.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { hashLoginInfo, LoginInfo } from "@/models/login";
+import { hashLoginInfo, LoginInfo } from "@/models/dashboard/login";
 
 const router = express.Router();
