+++ backend/src/services/sdk/session.ts
@@ -1,5 +1,5 @@
 import { Session } from "@prisma/client";
-import { createSession } from "@/models/session";
+import { createSession } from "@/models/sdk/session";
 
 export const handleNewSession = async ({
   createSessionFn,
